refactor: move restart_agent_in_container to cella-backend
