+++ crates/cella-backend/src/agent.rs
@@ -1,3 +1,8 @@
+use tracing::{info, warn};
+
+use crate::traits::ContainerBackend;
+use crate::types::ExecOptions;
+
 /// Environment variables injected when the managed cella-agent is available.
 pub fn agent_env_vars() -> Vec<String> {
     let version = env!("CARGO_PKG_VERSION");
@@ -5,6 +10,33 @@ pub fn agent_env_vars() -> Vec<String> {
         "BROWSER=/cella/bin/cella-browser".to_string(),
         format!("CELLA_AGENT_VERSION={version}"),
     ]
+}
+
+pub async fn restart_agent_in_container(client: &dyn ContainerBackend, container_id: &str) {
+    let agent_path = "/cella/bin/cella-agent";
+    let script = format!(
+        "pkill -f 'cella-agent daemon' 2>/dev/null; \
+         sleep 1; \
+         pgrep -f '[c]ella-agent daemon' >/dev/null 2>&1 || \
+         \"{agent_path}\" daemon \
+         --poll-interval \"${{CELLA_PORT_POLL_INTERVAL:-1000}}\" &"
+    );
+
+    match client
+        .exec_detached(
+            container_id,
+            &ExecOptions {
+                cmd: vec!["sh".to_string(), "-c".to_string(), script],
+                user: Some("root".to_string()),
+                env: None,
+                working_dir: None,
+            },
+        )
+        .await
+    {
+        Ok(_) => info!("Agent restart triggered in container {container_id}"),
+        Err(e) => warn!("Failed to restart agent in container: {e}"),
+    }
 }
 
 #[cfg(test)]
